+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 - Fix set typeFace without custom font
 - Improve rendering performance for messages without Markdown markup
-- Added Attachment Border Color attrs
+- Add Attachment Border Color attrs
+- Add Message Link Text Color attrs
 
 ## November 28th, 2019 - 3.4.1